+++ CHANGELOG.md
@@ -7,3 +7,10 @@
 - Added some unit tests for a few functions.
 - Added Travis CI
 - Added Procfile for easy deployment to Heroku
+
+## Version 0.0.2
+- Fixed standings bug and temporarily fixed the scores bug.
+- Updated to version 1.0.6 of sleeper-wrapper
+
+## Version 0.0.3
+- Fixed bug caused by KeyError in sleeper-wrapper